+++ commonroad/scenario/traffic_sign_interpreter.py
@@ -8,7 +8,7 @@ from commonroad.scenario.lanelet import LaneletNetwork
 __author__ = "Sebastian Maierhofer"
 __copyright__ = "TUM Cyber-Physical Systems Group"
 __credits__ = ["CAR@TUM"]
-__version__ = "2021.2"
+__version__ = "2021.3"
 __maintainer__ = "Sebastian Maierhofer"
 __email__ = "commonroad@lists.lrz.de"
 __status__ = "Release"