+++ commonroad/tests/geometry/test_shape.py
@@ -5,7 +5,7 @@ from commonroad.geometry.shape import Rectangle, Circle, Polygon, ShapeGroup
 __author__ = "Moritz Untersperger"
 __copyright__ = "TUM Cyber-Physical Systems Group"
 __credits__ = ["Priority Program SPP 1835 Cooperative Interacting Automobiles"]
-__version__ = "2021.2"
+__version__ = "2021.3"
 __maintainer__ = "Moritz Untersperger"
 __email__ = "commonroad@lists.lrz.de"
 __status__ = "Released"